feat : 예외 테스트 클래스 분리
